feat : added grpc communication between user and product service (adding product to cart api)

+++ server/Product_service/domain_product/product.go
@@ -13,7 +13,7 @@ type Product struct {
 	ProductImageSrc string
 	Name            string
 	Description     string
-	Price           int
+	Price           float32
 	Unit            int
 	Available       bool
 	Category        string
